Vereinfachungen in js und html
