+++ app/src/main/assets/src/html/onsPages/listview.html
@@ -1,4 +1,4 @@
-<ons-page id="page1">
+<ons-page id="listview">
   <ons-toolbar>
     <div class="center">Passwortübersicht</div>
     <div class="right ">
@@ -20,10 +20,7 @@
       <ons-list-item modifier="tappable">
         Benutzer Löschen
       </ons-list-item>
-      <ons-list-item modifier="tappable">
-        Neues Passwort
-      </ons-list-item>
-      <ons-list-item modifier="tappable">
+      <ons-list-item modifier="tappable" onclick="history.back()">
         Abmelden
       </ons-list-item>
     </ons-list>
