--- app/src/main/assets/src/html/onsPages/page2.html
@@ -1,33 +1,0 @@
-<!--not used-->
-<ons-page id="page2">
-    <ons-toolbar>
-        <div class="left">
-            <ons-back-button>Zurück zu Liste</ons-back-button>
-        </div>
-        <div class="center"></div>
-    </ons-toolbar>
-
-    <ons-list>
-        <ons-list-item expandable id="website2">
-            item
-            <div class="expandable-content">Expandable content</div>
-        </ons-list-item>
-        <ons-list-item id="username2">user</ons-list-item>
-        <ons-list-item id="password2">password</ons-list-item>
-    </ons-list>
-
-    <div style="text-align: center; margin-top: 30px;">
-        <p>
-            <ons-input id="website" modifier="underbar" float></ons-input>
-        </p>
-        <p>
-            <ons-input id="username" modifier="underbar" float></ons-input>
-        </p>
-        <p>
-            <ons-input id="password" modifier="underbar" type="password" float></ons-input>
-        </p>
-
-        <ons-button onclick="Cancel()">Löschen</ons-button>
-        <ons-button onclick="Commit()">Speichern</ons-button>
-    </div>
-</ons-page>
