+++ app/src/main/assets/src/html/onsPages/passwordDetailed.html
@@ -5,8 +5,8 @@
         </div>
         <div class="center"></div>
         <div class="right">
-            <ons-toolbar-button id="moreButton">
-                <ons-icon icon="ion-md-more"></ons-icon>
+            <ons-toolbar-button id="deleteButton" modifier="tappable">
+                <ons-icon icon="ion-md-trash"></ons-icon>
             </ons-toolbar-button>
         </div>
     </ons-toolbar>
@@ -34,25 +34,5 @@
             <ons-button id="abortButton" modifier="quiet" style="display: none">Abbrechen</ons-button>
             <ons-button id="commitButton" style="display: none">Speichern</ons-button>
         </ons-list-item>
-        <ons-list-item id="deleteWrapper" style="padding-right: 8px; display: none">
-            <ons-button id="deleteButton" modifier="large" style="background-color: red">
-                <ons-icon icon="ion-md-trash" size="20px"></ons-icon>
-                Passwort Löschen
-            </ons-button>
-        </ons-list-item>
     </ons-list>
-
-    <ons-popover cancelable direction="down" id="popover">
-        <ons-list width: max-content>
-            <ons-list-item modifier="tappable">
-                Passwort Löschen
-            </ons-list-item>
-            <ons-list-item modifier="tappable">
-                Neues Passwort
-            </ons-list-item>
-            <ons-list-item modifier="tappable">
-                Abmelden
-            </ons-list-item>
-        </ons-list>
-    </ons-popover>
 </ons-page>
